+++ main/hardware/terminal.hh
@@ -517,6 +517,7 @@ public:
     int NextDialog(Zone *z);    // stores a dialog to pop up after the current dialog closes
     int KillDialog();           // closes open dialog
     int HomePage();             // returns home page #
+    int GetDefaultLoginPage();  // returns appropriate login page based on terminal type and user
     int UpdateAllTerms(int update_message, const genericChar* value);
     int UpdateOtherTerms(int update_message, const genericChar* value);
     int TermsInUse();           // # of terms in use (total)